int casting

+++ asreview2-optuna/main.py
@@ -86,7 +86,7 @@ def n_query_extreme(results, n_records):
 def run_parallel(studies, *args, **kwargs):
     losses = defaultdict(list)
     gains = defaultdict(list)
-    with ProcessPoolExecutor(max_workers=(mp.cpu_count() - 1)/4) as executor:
+    with ProcessPoolExecutor(max_workers=int((mp.cpu_count() - 1)/4)) as executor:
         # Submit tasks
         futures = {
             executor.submit(process_row, row, *args, **kwargs): i
